+++ src/clj/cascalog/workflow.clj
@@ -6,13 +6,14 @@
         [jackknife.core :only (safe-assert)]
         [jackknife.seq :only (collectify)])
   (:require [cascalog.conf :as conf]
+            [cascalog.vars :as v]
             [cascalog.util :as u]
             [hadoop-util.core :as hadoop])
   (:import [cascalog Util]
            [java.io File]
            [java.util ArrayList]
            [cascading.tuple Tuple TupleEntry Fields]
-           [cascading.scheme.hadoop TextLine SequenceFile]
+           [cascading.scheme.hadoop TextLine SequenceFile TextDelimited]
            [cascading.scheme Scheme]
            [cascading.tap Tap SinkMode]
            [cascading.tap.hadoop Hfs Lfs GlobHfs TemplateTap
@@ -469,6 +470,17 @@
 
 (defn sequence-file [field-names]
   (SequenceFile. (fields field-names)))
+
+(defn delimited
+  [field-seq delim & {:keys [classes skip-header?]}]
+  (let [skip-header? (boolean skip-header?)
+        field-seq    (fields field-seq)
+        field-seq    (if (and classes (not (.isDefined field-seq)))
+                       (fields (v/gen-nullable-vars (count classes)))
+                       field-seq)]
+    (if classes
+      (TextDelimited. field-seq skip-header? delim (into-array classes))
+      (TextDelimited. field-seq skip-header? delim))))
 
 (deffilterop equal [& objs]
   (apply = objs))